refactor: Moved mamba_selectors
